report_generator/requirements.txt
psycopg2-binary
requests
schedule
redis
python-dotenv